+++ pkgs/default.nix
@@ -19,7 +19,6 @@ let
 in
 {
   dovecot-scripts = pkgs.callPackage ./dovecot-scripts.nix { };
-  notmuch-scripts = pkgs.callPackage ./notmuch-scripts.nix { };
 
   sieve-user = pkgs.callPackage ./sieve-user.nix {
     inherit mkSieveDerivation;