+++ test/client/default.nix
@@ -26,11 +26,6 @@ pkgs.nixosTest {
 
       mailrc.enable = true;
       mailrc.sieve.enable = true;
-
-      mailrc.muchsync = {
-        enable = true;
-        remotes = [ "localhost" ];
-      };
     };
   };
 
@@ -46,9 +41,6 @@ pkgs.nixosTest {
     machine.succeed("test -L /home/${user.systemUser}/.config/dovecot/sieve/active")
 
     machine.wait_for_unit("dovecot2.service")
-    machine.succeed("test -L /var/lib/dovecot/pipe-bin/notmuch-insert.sh")
-
     machine.succeed("su - ${user.systemUser} -c mailrc-tests")
-    machine.succeed("test -e /home/${user.systemUser}/mail/.notmuch/xapian/position.glass")
   '';
 }